Fix typos and connections to scripts
Mainly just for Melee Enemy
