+++ Assets/Prefabs/Prototyping/MeleeEnemy.prefab
@@ -16,7 +16,7 @@ GameObject:
   - component: {fileID: 7604672835813857726}
   - component: {fileID: 7604672835813857727}
   m_Layer: 0
-  m_Name: MeeleEnemy
+  m_Name: MeleeEnemy
   m_TagString: Untagged
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
