+++ Assets/Scenes/CharController Playground.unity
@@ -276,6 +276,75 @@ PrefabInstance:
       objectReference: {fileID: 0}
     m_RemovedComponents: []
   m_SourcePrefab: {fileID: 100100000, guid: ae0aef9b8e57d6b46b59713827c41a51, type: 3}
+--- !u!1001 &181667465
+PrefabInstance:
+  m_ObjectHideFlags: 0
+  serializedVersion: 2
+  m_Modification:
+    m_TransformParent: {fileID: 0}
+    m_Modifications:
+    - target: {fileID: 2777880356655725006, guid: ae0aef9b8e57d6b46b59713827c41a51,
+        type: 3}
+      propertyPath: m_Name
+      value: UnitSpawnPoint
+      objectReference: {fileID: 0}
+    - target: {fileID: 2777880356655725007, guid: ae0aef9b8e57d6b46b59713827c41a51,
+        type: 3}
+      propertyPath: m_RootOrder
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 2777880356655725007, guid: ae0aef9b8e57d6b46b59713827c41a51,
+        type: 3}
+      propertyPath: m_LocalPosition.x
+      value: 1.1956801
+      objectReference: {fileID: 0}
+    - target: {fileID: 2777880356655725007, guid: ae0aef9b8e57d6b46b59713827c41a51,
+        type: 3}
+      propertyPath: m_LocalPosition.y
+      value: 0.0000009536743
+      objectReference: {fileID: 0}
+    - target: {fileID: 2777880356655725007, guid: ae0aef9b8e57d6b46b59713827c41a51,
+        type: 3}
+      propertyPath: m_LocalPosition.z
+      value: 9.676219
+      objectReference: {fileID: 0}
+    - target: {fileID: 2777880356655725007, guid: ae0aef9b8e57d6b46b59713827c41a51,
+        type: 3}
+      propertyPath: m_LocalRotation.w
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 2777880356655725007, guid: ae0aef9b8e57d6b46b59713827c41a51,
+        type: 3}
+      propertyPath: m_LocalRotation.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 2777880356655725007, guid: ae0aef9b8e57d6b46b59713827c41a51,
+        type: 3}
+      propertyPath: m_LocalRotation.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 2777880356655725007, guid: ae0aef9b8e57d6b46b59713827c41a51,
+        type: 3}
+      propertyPath: m_LocalRotation.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 2777880356655725007, guid: ae0aef9b8e57d6b46b59713827c41a51,
+        type: 3}
+      propertyPath: m_LocalEulerAnglesHint.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 2777880356655725007, guid: ae0aef9b8e57d6b46b59713827c41a51,
+        type: 3}
+      propertyPath: m_LocalEulerAnglesHint.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 2777880356655725007, guid: ae0aef9b8e57d6b46b59713827c41a51,
+        type: 3}
+      propertyPath: m_LocalEulerAnglesHint.z
+      value: 0
+      objectReference: {fileID: 0}
+    m_RemovedComponents: []
+  m_SourcePrefab: {fileID: 100100000, guid: ae0aef9b8e57d6b46b59713827c41a51, type: 3}
 --- !u!1001 &413457891
 PrefabInstance:
   m_ObjectHideFlags: 0
@@ -460,7 +529,7 @@ Transform:
   m_ConstrainProportionsScale: 0
   m_Children: []
   m_Father: {fileID: 0}
-  m_RootOrder: 7
+  m_RootOrder: 8
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!114 &479960123
 MonoBehaviour:
@@ -1088,7 +1157,7 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: b60a437d5d2260441ac5922af2dd2397, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  meeleEnemyPrefab: {fileID: 7604672835813857724, guid: 83e7d315051105842a28fc65a639dd60,
+  meleeEnemyPrefab: {fileID: 7604672835813857724, guid: 83e7d315051105842a28fc65a639dd60,
     type: 3}
   hpRange: 0100000003000000
 --- !u!4 &930430971
@@ -1104,7 +1173,7 @@ Transform:
   m_ConstrainProportionsScale: 0
   m_Children: []
   m_Father: {fileID: 0}
-  m_RootOrder: 1
+  m_RootOrder: 2
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!1 &946466057
 GameObject:
@@ -1135,7 +1204,7 @@ Transform:
   m_ConstrainProportionsScale: 0
   m_Children: []
   m_Father: {fileID: 0}
-  m_RootOrder: 2
+  m_RootOrder: 3
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!1 &1028513131
 GameObject:
@@ -1283,7 +1352,7 @@ Mesh:
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_Name: pb_Mesh-1994
+  m_Name: pb_Mesh-28220
   serializedVersion: 10
   m_SubMeshes:
   - serializedVersion: 2
@@ -1539,7 +1608,7 @@ Transform:
   m_ConstrainProportionsScale: 0
   m_Children: []
   m_Father: {fileID: 0}
-  m_RootOrder: 4
+  m_RootOrder: 5
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!4 &1243973166 stripped
 Transform:
@@ -2109,7 +2178,7 @@ PrefabInstance:
     - target: {fileID: 2315940580372054155, guid: ab9cabdbbce019942a0183ce34f18497,
         type: 3}
       propertyPath: m_RootOrder
-      value: 3
+      value: 4
       objectReference: {fileID: 0}
     - target: {fileID: 2315940580372054155, guid: ab9cabdbbce019942a0183ce34f18497,
         type: 3}
@@ -2639,7 +2708,7 @@ Transform:
   - {fileID: 542508512}
   - {fileID: 874605736}
   m_Father: {fileID: 0}
-  m_RootOrder: 6
+  m_RootOrder: 7
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!1 &2114283754330233424
 GameObject:
@@ -6581,7 +6650,7 @@ Transform:
   - {fileID: 8379900264020877776}
   - {fileID: 1389714430}
   m_Father: {fileID: 0}
-  m_RootOrder: 5
+  m_RootOrder: 6
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!1 &8345070699414287371
 GameObject:
